+++ spec/integration/braintree/transaction_spec.rb
@@ -2280,6 +2280,37 @@ describe Braintree::Transaction do
         line_item.total_amount.should == BigDecimal.new("45.15")
       end
 
+      it "allows creation with single line item with zero amount fields and returns it" do
+        result = Braintree::Transaction.create(
+          :type => "sale",
+          :amount => "45.15",
+          :payment_method_nonce => Braintree::Test::Nonce::Transactable,
+          :line_items => [
+            {
+              :quantity => "1.0232",
+              :name => "Name #1",
+              :kind => "debit",
+              :unit_amount => "45.1232",
+              :total_amount => "45.15",
+              :unit_tax_amount => "0",
+              :discount_amount => "0",
+              :tax_amount => "0",
+            },
+          ],
+        )
+        result.success?.should == true
+        result.transaction.line_items.length.should == 1
+        line_item = result.transaction.line_items[0]
+        line_item.quantity.should == BigDecimal.new("1.0232")
+        line_item.name.should == "Name #1"
+        line_item.kind.should == "debit"
+        line_item.unit_amount.should == BigDecimal.new("45.1232")
+        line_item.total_amount.should == BigDecimal.new("45.15")
+        line_item.unit_tax_amount.should == BigDecimal.new("0")
+        line_item.discount_amount.should == BigDecimal.new("0")
+        line_item.tax_amount.should == BigDecimal.new("0")
+      end
+
       it "allows creation with single line item and returns it" do
         result = Braintree::Transaction.create(
           :type => "sale",
@@ -2513,7 +2544,7 @@ describe Braintree::Transaction do
         result.errors.for(:transaction).for(:line_items).for(:index_1).on(:discount_amount)[0].code.should == Braintree::ErrorCodes::TransactionLineItem::DiscountAmountIsTooLarge
       end
 
-      it "handles validation error discount amount must be greater than zero" do
+      it "handles validation error discount amount cannot be negative" do
         result = Braintree::Transaction.create(
           :type => "sale",
           :amount => "35.05",
@@ -2544,7 +2575,7 @@ describe Braintree::Transaction do
           ],
         )
         result.success?.should == false
-        result.errors.for(:transaction).for(:line_items).for(:index_1).on(:discount_amount)[0].code.should == Braintree::ErrorCodes::TransactionLineItem::DiscountAmountMustBeGreaterThanZero
+        result.errors.for(:transaction).for(:line_items).for(:index_1).on(:discount_amount)[0].code.should == Braintree::ErrorCodes::TransactionLineItem::DiscountAmountCannotBeNegative
       end
 
       it "handles validation error tax amount format is invalid" do
@@ -2595,7 +2626,7 @@ describe Braintree::Transaction do
         result.errors.for(:transaction).for(:line_items).for(:index_0).on(:tax_amount)[0].code.should == Braintree::ErrorCodes::TransactionLineItem::TaxAmountIsTooLarge
       end
 
-      it "handles validation error tax amount must be greater than zero" do
+      it "handles validation error tax amount cannot be negative" do
         result = Braintree::Transaction.create(
           :type => "sale",
           :amount => "35.05",
@@ -2608,7 +2639,7 @@ describe Braintree::Transaction do
               :unit_amount => "45.1232",
               :unit_of_measure => "gallon",
               :discount_amount => "1.02",
-              :tax_amount => "0",
+              :tax_amount => "-2",
               :total_amount => "45.15",
               :product_code => "23434",
               :commodity_code => "9SAASSD8724",
@@ -2616,7 +2647,7 @@ describe Braintree::Transaction do
           ],
         )
         result.success?.should == false
-        result.errors.for(:transaction).for(:line_items).for(:index_0).on(:tax_amount)[0].code.should == Braintree::ErrorCodes::TransactionLineItem::TaxAmountMustBeGreaterThanZero
+        result.errors.for(:transaction).for(:line_items).for(:index_0).on(:tax_amount)[0].code.should == Braintree::ErrorCodes::TransactionLineItem::TaxAmountCannotBeNegative
       end
 
       it "handles validation error kind is invalid" do
@@ -3263,7 +3294,7 @@ describe Braintree::Transaction do
         result.errors.for(:transaction).for(:line_items).for(:index_1).on(:unit_tax_amount)[0].code.should == Braintree::ErrorCodes::TransactionLineItem::UnitTaxAmountIsTooLarge
       end
 
-      it "handles validation error unit tax amount must be greater than zero" do
+      it "handles validation error unit tax amount cannot be negative" do
         result = Braintree::Transaction.create(
           :type => "sale",
           :amount => "35.05",
@@ -3295,7 +3326,7 @@ describe Braintree::Transaction do
           ],
         )
         result.success?.should == false
-        result.errors.for(:transaction).for(:line_items).for(:index_1).on(:unit_tax_amount)[0].code.should == Braintree::ErrorCodes::TransactionLineItem::UnitTaxAmountMustBeGreaterThanZero
+        result.errors.for(:transaction).for(:line_items).for(:index_1).on(:unit_tax_amount)[0].code.should == Braintree::ErrorCodes::TransactionLineItem::UnitTaxAmountCannotBeNegative
       end
 
       it "handles validation errors on line items structure" do